requirements.txt
Flask==2.3.3
psycopg2==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0